Small but major bug fixes

+++ src/common/FileReader.cpp
@@ -389,11 +389,11 @@ BspFile *FileReader::ReadFile(std::string fileName) {
 		readLumel.faceIndex = stoi(line);
 
 		std::getline(readFile, line);
-		readLumel.color[0] = stoi(line);
+		readLumel.color[0] = stof(line);
 		std::getline(readFile, line);
-		readLumel.color[1] = stoi(line);
+		readLumel.color[1] = stof(line);
 		std::getline(readFile, line);
-		readLumel.color[2] = stoi(line);
+		readLumel.color[2] = stof(line);
 
 		bspFile->fileLightmaps[i] = readLumel;
 	}
